Comment explainin data representation added, loading process fixed.
Also ELANDocumentJquery renamed.
And let's visit again zooming/ws plugins bugs with more precise
description:
1) If zooming was called at least once, plugins will render only part
with width equal to visible window width. It means that if you see, for
example, 1000pxs, but fullWSWidth is 15000pxs, timeline will be drawn
only on 1000pxs. I am not sure whether it is a bug in the plugin or not,
since it is a peculiar usage of it.
2) If plugins were turned on and off (destroyed), zooming will fail. I
have filed an issue on that:
https://github.com/katspaugh/wavesurfer.js/issues/848
These two were present in the older versions too, before changing
dict->array data representation.
3) No tab of the table with annotations is selected after zooming.
This is a new bug.